+++ modules/yield-optimization/yieldOptimizationService.test.ts
@@ -28,11 +28,8 @@ describe('YieldOptimizationService', () => {
   const sandbox = Sinon.createSandbox();
   const fetchStub = sandbox.stub(dom.window, 'fetch');
 
-  const createService = (
-    config: YieldOptimizationConfig,
-    labels: string[] = []
-  ): YieldOptimizationService => {
-    return new YieldOptimizationService(config, labels, noopLogger, domWindow);
+  const createService = (config: YieldOptimizationConfig): YieldOptimizationService => {
+    return new YieldOptimizationService(config, noopLogger, domWindow);
   };
 
   afterEach(() => {
@@ -43,21 +40,21 @@ describe('YieldOptimizationService', () => {
     const config: INoYieldOptimizationConfig = { provider: 'none' };
     const service = createService(config);
 
-    it('should always return undefined', () => {
-      return service.getPriceRule('foo').then(rule => {
-        expect(rule).to.be.undefined;
-      });
+    it('should always return undefined', async () => {
+      await service.init([]);
+      const rule = await service.getPriceRule('foo');
+      expect(rule).to.be.undefined;
     });
 
-    it('should not set any key-values', () => {
+    it('should not set any key-values', async () => {
       const adSlot = googleAdSlotStub('/123/publisher/p_content_1', 'ad_content_1');
       const getSlotElementIdSpy = sandbox.spy(adSlot, 'getSlotElementId');
       const setTargetingSpy = sandbox.spy(adSlot, 'setTargeting');
 
-      return service.setTargeting(adSlot).then(_ => {
-        expect(getSlotElementIdSpy).to.have.been.calledOnce;
-        expect(setTargetingSpy).to.not.have.been.called;
-      });
+      await service.init([]);
+      await service.setTargeting(adSlot);
+      expect(getSlotElementIdSpy).to.have.been.calledOnce;
+      expect(setTargetingSpy).to.not.have.been.called;
     });
   });
 
@@ -66,21 +63,21 @@ describe('YieldOptimizationService', () => {
       const config: IStaticYieldOptimizationConfig = { provider: 'static', config: { rules: {} } };
       const service = createService(config);
 
-      it('should always return undefined', () => {
-        return service.getPriceRule('foo').then(rule => {
-          expect(rule).to.be.undefined;
-        });
+      it('should always return undefined', async () => {
+        await service.init([]);
+        const rule = await service.getPriceRule('foo');
+        expect(rule).to.be.undefined;
       });
 
-      it('should not set any key-values', () => {
+      it('should not set any key-values', async () => {
         const adSlot = googleAdSlotStub('/123/publisher/p_content_1', 'ad_content_1');
         const getSlotElementIdSpy = sandbox.spy(adSlot, 'getSlotElementId');
         const setTargetingSpy = sandbox.spy(adSlot, 'setTargeting');
 
-        return service.setTargeting(adSlot).then(_ => {
-          expect(getSlotElementIdSpy).to.have.been.calledOnce;
-          expect(setTargetingSpy).to.not.have.been.called;
-        });
+        await service.init([]);
+        await service.setTargeting(adSlot);
+        expect(getSlotElementIdSpy).to.have.been.calledOnce;
+        expect(setTargetingSpy).to.not.have.been.called;
       });
     });
 
@@ -99,22 +96,22 @@ describe('YieldOptimizationService', () => {
         }
       };
 
-      it('should return undefined if the ad slot is not configured', () => {
+      it('should return undefined if the ad slot is not configured', async () => {
         const service = createService(config);
-        return service.getPriceRule('foo').then(rule => {
-          expect(rule).to.be.undefined;
-        });
+        await service.init([]);
+        const rule = await service.getPriceRule('foo');
+        expect(rule).to.be.undefined;
       });
 
-      it('should return the price rule', () => {
-        const service = createService(config, []);
-        return service.getPriceRule(adUnit).then(rule => {
-          expect(rule).to.be.ok;
-          expect(rule!.main).to.be.true;
-          expect(rule!.priceRuleId).to.be.eq(3);
-          expect(rule!.floorprice).to.be.eq(0.2);
-          expect(rule!.main).to.be.eq(true);
-        });
+      it('should return the price rule', async () => {
+        const service = createService(config);
+        await service.init([]);
+        const rule = await service.getPriceRule(adUnit);
+        expect(rule).to.be.ok;
+        expect(rule!.main).to.be.true;
+        expect(rule!.priceRuleId).to.be.eq(3);
+        expect(rule!.floorprice).to.be.eq(0.2);
+        expect(rule!.main).to.be.eq(true);
       });
     });
   });
@@ -137,7 +134,7 @@ describe('YieldOptimizationService', () => {
     };
 
     describe('error handling', () => {
-      it('should retry three times', () => {
+      it('should retry three times', async () => {
         fetchStub.onFirstCall().rejects('FetchRequestFailed');
         fetchStub.onSecondCall().rejects('FetchRequestFailed');
         fetchStub.onThirdCall().resolves({
@@ -147,24 +144,24 @@ describe('YieldOptimizationService', () => {
           }
         } as any);
         const service = createService(config);
+        await service.init([]);
 
-        return service.getPriceRule(adUnit).then(rule => {
-          expect(rule).not.to.be.undefined;
-          expect(fetchStub).to.have.been.calledThrice;
-        });
+        const rule = await service.getPriceRule(adUnit);
+        expect(rule).not.to.be.undefined;
+        expect(fetchStub).to.have.been.calledThrice;
       });
 
-      it('should always return undefined if the asset loading fails after 3 retries', () => {
+      it('should always return undefined if the asset loading fails after 3 retries', async () => {
         fetchStub.rejects('FetchRequestFailed');
         const service = createService(config);
 
-        return service.getPriceRule(adUnit).then(rule => {
-          expect(rule).to.be.undefined;
-          expect(fetchStub).to.have.been.calledThrice;
-        });
+        await service.init([]);
+        const rule = await service.getPriceRule(adUnit);
+        expect(rule).to.be.undefined;
+        expect(fetchStub).to.have.been.calledThrice;
       });
 
-      it('should always return undefined if the response is not json', () => {
+      it('should always return undefined if the response is not json', async () => {
         fetchStub.resolves({
           ok: true,
           json: (): Promise<any> => {
@@ -172,14 +169,14 @@ describe('YieldOptimizationService', () => {
           }
         } as any);
         const service = createService(config);
+        await service.init([]);
 
-        return service.getPriceRule(adUnit).then(rule => {
-          expect(rule).to.be.undefined;
-          expect(fetchStub).to.have.been.called;
-        });
+        const rule = await service.getPriceRule(adUnit);
+        expect(rule).to.be.undefined;
+        expect(fetchStub).to.have.been.called;
       });
 
-      it('should always return undefined if the config json has the wrong format', () => {
+      it('should always return undefined if the config json has the wrong format', async () => {
         fetchStub.resolves({
           ok: true,
           json: (): Promise<any> => {
@@ -187,11 +184,11 @@ describe('YieldOptimizationService', () => {
           }
         } as any);
         const service = createService(config);
+        await service.init([]);
 
-        return service.getPriceRule(adUnit).then(rule => {
-          expect(rule).to.be.undefined;
-          expect(fetchStub).to.have.been.called;
-        });
+        const rule = await service.getPriceRule(adUnit);
+        expect(rule).to.be.undefined;
+        expect(fetchStub).to.have.been.called;
       });
     });
 
@@ -205,36 +202,74 @@ describe('YieldOptimizationService', () => {
         } as any);
       });
 
-      it('should call the configured endpoint only once', () => {
+      it('should call the configured endpoint only once', async () => {
         const service = createService(config);
         const adSlot = googleAdSlotStub('/123/publisher/p_content_1', adUnit);
 
-        return Promise.all([service.getPriceRule(adUnit), service.setTargeting(adSlot)]).then(_ => {
+        await service.init([]);
+        await service.getPriceRule(adUnit);
+        await service.setTargeting(adSlot);
+
+        expect(fetchStub).to.have.been.calledOnce;
+        expect(fetchStub).to.have.been.calledOnceWithExactly(config.configEndpoint, {
+          body: '{"device":"mobile"}',
+          headers: { Accept: 'application/json', 'Content-Type': 'application/json' },
+          method: 'POST',
+          mode: 'cors'
+        });
+      });
+
+      it('should return undefined if the ad slot is not configured', async () => {
+        const service = createService(config);
+        await service.init([]);
+        const rule = await service.getPriceRule('foo');
+        expect(rule).to.be.undefined;
+      });
+
+      it('should return the price rule', async () => {
+        const service = createService(config);
+        await service.init([]);
+        const rule = await service.getPriceRule(adUnit);
+        expect(rule).to.be.ok;
+        expect(rule!.main).to.be.true;
+        expect(rule!.priceRuleId).to.be.eq(3);
+        expect(rule!.floorprice).to.be.eq(0.2);
+        expect(rule!.main).to.be.eq(true);
+      });
+    });
+
+    describe('init', () => {
+      beforeEach(() => {
+        fetchStub.resolves({
+          ok: true,
+          json: (): Promise<any> => {
+            return Promise.resolve(publisherYieldConfiguration);
+          }
+        } as any);
+      });
+
+      // input parameters
+      [
+        { labels: [], label: 'mobile' },
+        { labels: ['mobile'], label: 'mobile' },
+        { labels: ['mobile', 'desktop'], label: 'desktop' },
+        { labels: ['desktop'], label: 'desktop' }
+      ].forEach(({ labels, label }) => {
+        it(`should use ${label} with input [${labels.join(',')}]`, async () => {
+          const service = createService(config);
+          const adSlot = googleAdSlotStub('/123/publisher/p_content_1', adUnit);
+
+          await service.init(labels);
+          await service.getPriceRule(adUnit);
+          await service.setTargeting(adSlot);
+
           expect(fetchStub).to.have.been.calledOnce;
           expect(fetchStub).to.have.been.calledOnceWithExactly(config.configEndpoint, {
-            body: '{"device":"mobile"}',
+            body: `{"device":"${label}"}`,
             headers: { Accept: 'application/json', 'Content-Type': 'application/json' },
             method: 'POST',
             mode: 'cors'
           });
-        });
-      });
-
-      it('should return undefined if the ad slot is not configured', () => {
-        const service = createService(config);
-        return service.getPriceRule('foo').then(rule => {
-          expect(rule).to.be.undefined;
-        });
-      });
-
-      it('should return the price rule', () => {
-        const service = createService(config, []);
-        return service.getPriceRule(adUnit).then(rule => {
-          expect(rule).to.be.ok;
-          expect(rule!.main).to.be.true;
-          expect(rule!.priceRuleId).to.be.eq(3);
-          expect(rule!.floorprice).to.be.eq(0.2);
-          expect(rule!.main).to.be.eq(true);
         });
       });
     });