+++ CHANGELOG.md
@@ -2,9 +2,15 @@
 
 ## Unreleased
 
+## 3.13.1
+
+[GD-2313](https://jira.gutefrage.net/browse/GD-2313) Fix label configuration
+
 ## 3.13.0
 
 [GD-2313](https://jira.gutefrage.net/browse/GD-2313) Yield Optimization is now a module and needs to be configured as such.
+
+See [module README](modules/yield-optimization/README.md);
 
 ## 3.12.2